adjust token.pojo package structure
