+++ sharding-core/sharding-core-rewrite/src/main/java/org/apache/shardingsphere/core/rewrite/feature/sharding/token/generator/impl/TableTokenGenerator.java
@@ -33,7 +33,7 @@ import org.apache.shardingsphere.core.parse.sql.statement.SQLStatement;
 import org.apache.shardingsphere.core.parse.sql.statement.dml.SelectStatement;
 import org.apache.shardingsphere.core.rewrite.sql.token.generator.CollectionSQLTokenGenerator;
 import org.apache.shardingsphere.core.rewrite.feature.sharding.token.generator.ShardingRuleAware;
-import org.apache.shardingsphere.core.rewrite.sql.token.pojo.impl.TableToken;
+import org.apache.shardingsphere.core.rewrite.feature.sharding.token.pojo.TableToken;
 import org.apache.shardingsphere.core.rule.ShardingRule;
 
 import java.util.Collection;
