+++ sharding-core/sharding-core-rewrite/src/main/java/org/apache/shardingsphere/core/rewrite/sql/token/pojo/generic/RemoveToken.java
@@ -15,39 +15,30 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.core.rewrite.sql.token.pojo.impl;
+package org.apache.shardingsphere.core.rewrite.sql.token.pojo.generic;
 
 import lombok.Getter;
 import org.apache.shardingsphere.core.rewrite.sql.token.pojo.SQLToken;
 import org.apache.shardingsphere.core.rewrite.sql.token.pojo.Substitutable;
 
 /**
- * Insert cipher item token.
+ * Remove token.
  *
+ * @author zhangliang
  * @author panjuan
  */
-public final class SelectEncryptItemToken extends SQLToken implements Substitutable {
+@Getter
+public final class RemoveToken extends SQLToken implements Substitutable {
     
-    @Getter
     private final int stopIndex;
     
-    private final String columnName;
-    
-    private final String owner;
-    
-    public SelectEncryptItemToken(final int startIndex, final int stopIndex, final String columnName, final String owner) {
+    public RemoveToken(final int startIndex, final int stopIndex) {
         super(startIndex);
         this.stopIndex = stopIndex;
-        this.columnName = columnName;
-        this.owner = owner;
-    }
-    
-    public SelectEncryptItemToken(final int startIndex, final int stopIndex, final String columnName) {
-        this(startIndex, stopIndex, columnName, null);
     }
     
     @Override
     public String toString() {
-        return null == owner ? columnName : String.format("%s.%s", owner, columnName);
+        return "";
     }
 }
